Redirecting to HTTP
Recently GitHub has introduced HTTPS to GitHub Pages. Unfortunately, this breaks the setup tool, since it makes HTTP requests to the created droplet, in order to determine if the setup has completed. These requests are blocked in most browsers when the page is loaded over HTTPS. Since I want the whole procedure to run purely on the client-side, I added this code to ensure the page is always loaded over HTTP. Don't worry though, requests to the DigitalOcean API are always performed over HTTPS and are therefore secure.